lasers: move rendering into a new file
