+++ src/laser.h
@@ -15,23 +15,12 @@
 #include "projectile.h"
 #include "resource/shader_program.h"
 #include "entity.h"
+#include "laserdraw.h"
 
 typedef LIST_ANCHOR(Laser) LaserList;
 
 typedef cmplx (*LaserPosRule)(Laser* l, float time);
 typedef void (*LaserLogicRule)(Laser* l, int time);
-
-typedef struct LaserRenderData LaserRenderData;
-struct LaserRenderData {
-	// NOTE: this struct is considered private
-
-	LIST_INTERFACE(LaserRenderData);
-	int tile;
-
-	struct {
-		FloatOffset top_left, bottom_right;
-	} bbox;
-};
 
 DEFINE_ENTITY_TYPE(Laser, {
 	cmplx pos;
@@ -43,6 +32,7 @@ DEFINE_ENTITY_TYPE(Laser, {
 	LaserPosRule prule;
 	LaserLogicRule lrule;
 
+	// NOTE: this field is considered private
 	LaserRenderData _renderdata;
 
 	Color color;
@@ -66,9 +56,6 @@ DEFINE_ENTITY_TYPE(Laser, {
 #define create_lasercurve2c(p, time, deathtime, clr, rule, a0, a1) create_laser(p, time, deathtime, clr, rule, 0, a0, a1, 0, 0)
 #define create_lasercurve3c(p, time, deathtime, clr, rule, a0, a1, a2) create_laser(p, time, deathtime, clr, rule, 0, a0, a1, a2, 0)
 #define create_lasercurve4c(p, time, deathtime, clr, rule, a0, a1, a2, a3) create_laser(p, time, deathtime, clr, rule, 0, a0, a1, a2, a3)
-
-void lasers_preload(void);
-void lasers_free(void);
 
 void delete_lasers(void);
 void process_lasers(void);
